e2e test for magic
